refactor: chat에서 socket연결 로직 분리
